DBet: bf16 NaN guards (skip non-finite steps, fp32 CE/BCE, lower 1-GPU lr)
Loss/grad_norm went NaN ~87k steps once lr reached 1e-4: bf16 drafter (no fp32 master
on 1 GPU) overflows on a grad spike, then optimizer.step() poisons the weights
permanently. Fixes:
- train_dbet: skip_nonfinite_steps (default on) -- skip optimizer.step when grad_norm is
  NaN/Inf, zero grads, advance schedule; counts/logs skips. One overflow no longer kills
  the run.
- dbet_train_core: compute token CE (logits.float()) and confidence BCE (conf.float()) in
  fp32 -- bf16 logsumexp/log can overflow.
- dbet_bd_1gpu.yaml: lr 1e-4 -> 5e-5 (bf16-stable); quick-figure recipe lr 3e-4 -> 1e-4.
For a higher-lr fp32-master run, use the multi-GPU FSDP dbet_bd.yaml.

+++ dFactory/tasks/dbet_train_core.py
@@ -102,14 +102,15 @@ def dbet_train_step(model, micro_batch, n_micro_batches, args, mask_id=MASK_ID, 
     # decayed CE + confidence BCE on the remaining-masked positions vs golden
     w = decay_weights(remaining, bs)
     denom = w.sum().clamp_min(1.0)
-    ce = F.cross_entropy(logits.reshape(-1, logits.shape[-1]), clean_ids.reshape(-1),
+    # CE in fp32: bf16 logsumexp can overflow for large drafter logits (NaN). logits.float() is the stable path.
+    ce = F.cross_entropy(logits.reshape(-1, logits.shape[-1]).float(), clean_ids.reshape(-1),
                          reduction="none").view_as(clean_ids)
     tok_loss = (ce * w).sum() / denom
     loss = tok_loss
     metrics = {"tok": float(tok_loss.detach()), "n_remaining": int(remaining.sum())}
     if conf is not None:
         accept = (logits.argmax(-1) == clean_ids).float()             # label 1 iff drafter argmax == golden
-        c = conf.clamp(1e-5, 1 - 1e-5)
+        c = conf.float().clamp(1e-5, 1 - 1e-5)                         # fp32 for a stable log
         bce = -(accept * c.log() + (1 - accept) * (1 - c).log())
         conf_loss = (bce * w).sum() / denom
         loss = loss + args.train.conf_loss_weight * conf_loss
